KOGITO-9416 - Remove SVG extension from prod profile
