+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-common/runtime/pom.xml
@@ -21,7 +21,7 @@
         </dependency>
         <dependency>
             <groupId>org.kie.kogito</groupId>
-            <artifactId>kogito-addons-quarkus-process-svg</artifactId>
+            <artifactId>kogito-addons-process-svg</artifactId>
         </dependency>
         <dependency>
             <groupId>org.kie.kogito</groupId>
