+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-infinispan/integration-tests-sw/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+kogito.persistence.type=infinispan
+
+quarkus.kogito.data-index.graphql.ui.always-include=true
