+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-inmemory/integration-tests-sw/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.kogito.data-index.graphql.ui.always-include=true
