+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-mongodb/integration-tests-sw/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+kogito.persistence.type=mongodb
+
+quarkus.kogito.data-index.graphql.ui.always-include=true
+# MongoDB
+quarkus.mongodb.database=kogito
