+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-postgresql/integration-tests-process/src/main/resources/hello.bpmn
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn2:definitions xmlns:bpmn2="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:bpsim="http://www.bpsim.org/schemas/1.0" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:drools="http://www.jboss.org/drools" xmlns:xsi="xsi" id="_hT2oIPJzEDuE8tm0q8uK_w" xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL BPMN20.xsd http://www.jboss.org/drools drools.xsd http://www.bpsim.org/schemas/1.0 bpsim.xsd http://www.omg.org/spec/DD/20100524/DC DC.xsd http://www.omg.org/spec/DD/20100524/DI DI.xsd " exporter="jBPM Process Modeler" exporterVersion="2.0" targetNamespace="http://www.omg.org/bpmn20">
+  <bpmn2:collaboration id="_4326C578-FDBD-4F8C-BAB8-5C9CFE867882" name="Default Collaboration">
+    <bpmn2:participant id="_B70F49F9-D9BE-4BCD-A095-2601433AD6F7" name="Pool Participant" processRef="hello"/>
+  </bpmn2:collaboration>
+  <bpmn2:process id="hello" drools:packageName="com.example" drools:version="1.0" drools:adHoc="false" name="hello" isExecutable="true" processType="Public">
+    <bpmn2:sequenceFlow id="_9FD38FF8-A530-41E7-8EA0-E1F71353CBB3" sourceRef="_C3F8BD7A-55E5-4513-8C0D-5576F8099167" targetRef="_B3241ACF-97BE-443B-A49F-964AB3DD006C"/>
+    <bpmn2:endEvent id="_B3241ACF-97BE-443B-A49F-964AB3DD006C">
+      <bpmn2:incoming>_9FD38FF8-A530-41E7-8EA0-E1F71353CBB3</bpmn2:incoming>
+    </bpmn2:endEvent>
+    <bpmn2:startEvent id="_C3F8BD7A-55E5-4513-8C0D-5576F8099167">
+      <bpmn2:outgoing>_9FD38FF8-A530-41E7-8EA0-E1F71353CBB3</bpmn2:outgoing>
+    </bpmn2:startEvent>
+  </bpmn2:process>
+  <bpmndi:BPMNDiagram>
+    <bpmndi:BPMNPlane bpmnElement="hello">
+      <bpmndi:BPMNShape id="shape__C3F8BD7A-55E5-4513-8C0D-5576F8099167" bpmnElement="_C3F8BD7A-55E5-4513-8C0D-5576F8099167">
+        <dc:Bounds height="56" width="56" x="321" y="287"/>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="shape__B3241ACF-97BE-443B-A49F-964AB3DD006C" bpmnElement="_B3241ACF-97BE-443B-A49F-964AB3DD006C">
+        <dc:Bounds height="56" width="56" x="457" y="287"/>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge id="edge_shape__C3F8BD7A-55E5-4513-8C0D-5576F8099167_to_shape__B3241ACF-97BE-443B-A49F-964AB3DD006C" bpmnElement="_9FD38FF8-A530-41E7-8EA0-E1F71353CBB3">
+        <di:waypoint x="349" y="315"/>
+        <di:waypoint x="485" y="315"/>
+      </bpmndi:BPMNEdge>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+  <bpmn2:relationship type="BPSimData">
+    <bpmn2:extensionElements>
+      <bpsim:BPSimData>
+        <bpsim:Scenario id="default" name="Simulationscenario">
+          <bpsim:ScenarioParameters/>
+          <bpsim:ElementParameters elementRef="_C3F8BD7A-55E5-4513-8C0D-5576F8099167">
+            <bpsim:TimeParameters>
+              <bpsim:ProcessingTime>
+                <bpsim:NormalDistribution mean="0" standardDeviation="0"/>
+              </bpsim:ProcessingTime>
+            </bpsim:TimeParameters>
+          </bpsim:ElementParameters>
+        </bpsim:Scenario>
+      </bpsim:BPSimData>
+    </bpmn2:extensionElements>
+    <bpmn2:source>_hT2oIPJzEDuE8tm0q8uK_w</bpmn2:source>
+    <bpmn2:target>_hT2oIPJzEDuE8tm0q8uK_w</bpmn2:target>
+  </bpmn2:relationship>
+</bpmn2:definitions>
