+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-postgresql/integration-tests-sw/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+kogito.persistence.type=jdbc
+quarkus.flyway.migrate-at-start=true
+
+quarkus.kogito.data-index.graphql.ui.always-include=true
+
+quarkus.datasource.db-kind=postgresql
