+++ data-index/kogito-addons-quarkus-data-index/kogito-addons-quarkus-data-index-postgresql/pom.xml
@@ -15,7 +15,21 @@
   <modules>
     <module>deployment</module>
     <module>runtime</module>
-    <module>integration-tests</module>
+    <module>integration-tests-sw</module>
   </modules>
+
+  <profiles>
+    <profile>
+      <id>default</id>
+      <activation>
+        <property>
+          <name>!productized</name>
+        </property>
+      </activation>
+      <modules>
+        <module>integration-tests-process</module>
+      </modules>
+    </profile>
+  </profiles>
 
 </project>
